manager: split entities and services out to their own files
